requirements.txt
flask
scikit-learn
pillow
numpy
scikit-image
joblib
hypothesis
pytest
matplotlib
